pyproject.toml
[project]
name = "laptop-sync"
version = "0.1.0"
description = "Mirror files from a Windows directory to a remote Linux host over SSH"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "rich-click>=1.9.7",
]
